Move source-specific functions out of __init__.py into their reader modules
build_nahar_levels_and_phixs() and nahar_level_schema move to
readnahardata.py beside the NaharEnergyLevel row class they are built
from. check_forbidden() moves to readqubdata.py, its only caller.
read_storey_2016_upsilondata() moves to a new readstoreydata.py,
matching the one-module-per-data-source layout of the other readers
(its only reference is the commented-out Storey Co II line in the
qub_cobalt handler branch, updated to the new path).

__init__.py keeps only the shared machinery: handler dispatch, the
cross-reader configuration parsers, phixs reduction, and the output
writers. match_hydrogenic_phixs() stays because it is a cross-handler
fallback rather than a reader for any one source's files.

No behavior change: cmfgen/qub fixture checksums and the Nahar handler
output are byte-identical; 22 tests pass.
